(a) + Added layout parameter to templates of Univesal Content Module (adding some features to news and carousel templates, e.g. columns for featured in carousel template, e.g. image/content placement options in news template), (b) also some parameter restucturing for easier/cleaner basic configuration of the module, (c) + Added default image (thumbnailed according to configuration) as CORE parameter of the module instead of adding this to individual templates

+++ admin/controllers/templates.raw.php
@@ -119,6 +119,10 @@ class FlexicontentControllerTemplates extends FlexicontentController
 		$db = JFactory::getDBO();
 		if ($ext_view=='module') {
 			$query = 'SELECT params FROM #__modules WHERE id = '.$ext_id;
+			// load english language file for 'mod_flexicontent' module then override with current language file
+			$module_name = basename(dirname($directory));
+			JFactory::getLanguage()->load($module_name, JPATH_SITE, 'en-GB', true);
+			JFactory::getLanguage()->load($module_name, JPATH_SITE, null, true);
 		} else if ($ext_view=='field') {
 			$query = 'SELECT attribs FROM #__flexicontent_fields WHERE id = '.$ext_id;
 		} else {
@@ -130,10 +134,16 @@ class FlexicontentControllerTemplates extends FlexicontentController
 		
 		$layoutpath = $path.DS.$layout_name.'.xml';
 		if (!file_exists($layoutpath)) {
-			echo !FLEXI_J16GE ? '<div style="font-size: 11px; color: gray; background-color: lightyellow; border: 1px solid lightgray; width: auto; padding: 4px 2%; margin: 1px 8px; height: auto;">' : '<p class="tip">';
-			echo ' Currently selected layout: <b>"'.$layout_name.'"</b> does not have layout specific parameters';
-			echo !FLEXI_J16GE ? '</div>' : '</p>';
-			exit;
+			if (file_exists($path.DS.'_fallback'.DS.'_fallback.xml')) {
+				$layoutpath = $path.DS.'_fallback'.DS.'_fallback.xml';
+				echo '<div class="fcsep_level3">Currently selected layout: <b>"'.$layout_name.'"</b> does not have a parameters XML file, using general defaults</div><div class="clear"></div>';
+			}
+			else {
+				echo !FLEXI_J16GE ? '<div style="font-size: 11px; color: gray; background-color: lightyellow; border: 1px solid lightgray; width: auto; padding: 4px 2%; margin: 1px 8px; height: auto;">' : '<p class="tip">';
+				echo ' Currently selected layout: <b>"'.$layout_name.'"</b> does not have layout specific parameters';
+				echo !FLEXI_J16GE ? '</div>' : '</p>';
+				exit;
+			}
 		}
 		
 		//Get data from the model
